First version of moving methods from client to backend (many tests failing)
