+++ ibis/backends/dask/client.py
@@ -1,20 +1,16 @@
 """The dask client implementation."""
 from functools import partial
-from typing import Dict, Mapping
+from typing import Dict
 
 import dask.dataframe as dd
 import dateutil.parser
 import numpy as np
 import pandas as pd
-import toolz
-from dask.base import DaskMethodsMixin
 from pandas.api.types import DatetimeTZDtype
 
-import ibis.common.exceptions as com
 import ibis.expr.datatypes as dt
 import ibis.expr.operations as ops
 import ibis.expr.schema as sch
-import ibis.expr.types as ir
 from ibis.backends.base import Client, Database
 from ibis.backends.pandas.client import (
     PANDAS_DATE_TYPES,
@@ -24,8 +20,6 @@ from ibis.backends.pandas.client import (
     ibis_dtype_to_pandas,
     ibis_schema_to_pandas,
 )
-
-from .core import execute_and_reset
 
 infer_dask_dtype = pd.api.types.infer_dtype
 
@@ -159,104 +153,4 @@ class DaskDatabase(Database):
 class DaskClient(Client):
     def __init__(self, backend, dictionary: Dict[str, dd.DataFrame]):
         self.backend = backend
-        self.database_class = backend.database_class
-        self.table_class = backend.table_class
-        self.dictionary = dictionary
-
-    def table(self, name: str, schema: sch.Schema = None) -> DaskTable:
-        df = self.dictionary[name]
-        schema = sch.infer(df, schema=schema)
-        return self.table_class(name, schema, self).to_expr()
-
-    def execute(
-        self,
-        query: ir.Expr,
-        params: Mapping[ir.Expr, object] = None,
-        limit: str = 'default',
-        **kwargs,
-    ):
-        if limit != 'default':
-            raise ValueError(
-                'limit parameter to execute is not yet implemented in the '
-                'dask backend'
-            )
-
-        if not isinstance(query, ir.Expr):
-            raise TypeError(
-                "`query` has type {!r}, expected ibis.expr.types.Expr".format(
-                    type(query).__name__
-                )
-            )
-
-        result = self.compile(query, params, **kwargs)
-        if isinstance(result, DaskMethodsMixin):
-            return result.compute()
-        else:
-            return result
-
-    def compile(
-        self, query: ir.Expr, params: Mapping[ir.Expr, object] = None, **kwargs
-    ):
-        """Compile `expr`.
-
-        Notes
-        -----
-        For the dask backend returns a dask graph that you can run ``.compute``
-        on to get a pandas object.
-
-        """
-        return execute_and_reset(query, params=params, **kwargs)
-
-    def database(self, name: str = None) -> DaskDatabase:
-        """Construct a database called `name`."""
-        return self.database_class(name, self)
-
-    def load_data(self, table_name: str, obj: dd.DataFrame, **kwargs):
-        """Load data from `obj` into `table_name`.
-
-        Parameters
-        ----------
-        table_name : str
-        obj : dask.dataframe.DataFrame
-
-        """
-        # kwargs is a catch all for any options required by other backends.
-        self.dictionary[table_name] = obj
-
-    def create_table(
-        self,
-        table_name: str,
-        obj: dd.DataFrame = None,
-        schema: sch.Schema = None,
-    ):
-        """Create a table."""
-        if obj is not None:
-            df = obj
-        elif schema is not None:
-            dtypes = ibis_schema_to_dask(schema)
-            df = schema.apply_to(
-                dd.from_pandas(
-                    pd.DataFrame(columns=list(map(toolz.first, dtypes))),
-                    npartitions=1,
-                )
-            )
-        else:
-            raise com.IbisError('Must pass expr or schema')
-
-        self.dictionary[table_name] = df
-
-    def get_schema(self, table_name: str, database: str = None) -> sch.Schema:
-        """Return a Schema object for the indicated table and database.
-
-        Parameters
-        ----------
-        table_name : str
-            May be fully qualified
-        database : str
-
-        Returns
-        -------
-        ibis.expr.schema.Schema
-
-        """
-        return sch.infer(self.dictionary[table_name])
+        self.backend.dictionary = dictionary
