+++ ibis/backends/hdf5/__init__.py
@@ -13,31 +13,7 @@ class HDFTable(ops.DatabaseTable):
 
 
 class HDFClient(FileClient):
-    def insert(
-        self, path, key, expr, format='table', data_columns=True, **kwargs
-    ):
-
-        path = self.root / path
-        data = execute(expr)
-        data.to_hdf(
-            str(path), key, format=format, data_columns=data_columns, **kwargs
-        )
-
-    def table(self, name, path=None):
-        if path is None:
-            path = self.root / f"{name}.{self.extension}"
-
-        if name not in self.list_tables(path):
-            raise AttributeError(name)
-
-        # get the schema
-        with pd.HDFStore(str(path), mode='r') as store:
-            df = store.select(name, start=0, stop=50)
-            schema = sch.infer(df)
-
-        t = self.table_class(name, schema, self).to_expr()
-        self.dictionary[name] = path
-        return t
+    pass
 
 
 class Backend(BaseFileBackend):
@@ -87,6 +63,32 @@ class Backend(BaseFileBackend):
             )
         databases = self._list_databases_dirs_or_files(path)
         return self._filter_with_like(databases, like)
+
+    def insert(
+        self, path, key, expr, format='table', data_columns=True, **kwargs
+    ):
+
+        path = self.root / path
+        data = execute(expr)
+        data.to_hdf(
+            str(path), key, format=format, data_columns=data_columns, **kwargs
+        )
+
+    def table(self, name, path=None):
+        if path is None:
+            path = self.root / f"{name}.{self.extension}"
+
+        if name not in self.list_tables(path):
+            raise AttributeError(name)
+
+        # get the schema
+        with pd.HDFStore(str(path), mode='r') as store:
+            df = store.select(name, start=0, stop=50)
+            schema = sch.infer(df)
+
+        t = self.table_class(name, schema, self).to_expr()
+        self.dictionary[name] = path
+        return t
 
 
 @execute_node.register(Backend.table_class, HDFClient)
