+++ ibis/backends/pandas/__init__.py
@@ -1,9 +1,19 @@
 import pandas as pd
+import toolz
 
+import ibis.common.exceptions as com
 import ibis.config
+import ibis.expr.schema as sch
+import ibis.expr.types as ir
 from ibis.backends.base import BaseBackend
 
-from .client import PandasClient, PandasDatabase, PandasTable
+from .client import (
+    PandasClient,
+    PandasDatabase,
+    PandasTable,
+    ibis_schema_to_pandas,
+)
+from .core import execute_and_reset
 from .execution import execute
 from .udf import udf  # noqa F401
 
@@ -74,6 +84,24 @@ class BasePandasBackend(BaseBackend):
             list(self.client.dictionary.keys()), like
         )
 
+    def table(self, name: str, schema: sch.Schema = None):
+        df = self.dictionary[name]
+        schema = sch.infer(df, schema=schema)
+        return self.table_class(name, schema, self).to_expr()
+
+    def database(self, name=None):
+        return self.database_class(name, self)
+
+    def load_data(self, table_name, obj, **kwargs):
+        # kwargs is a catch all for any options required by other backends.
+        self.dictionary[table_name] = obj
+
+    def get_schema(self, table_name, database=None):
+        return sch.infer(self.dictionary[table_name])
+
+    def compile(self, expr, *args, **kwargs):
+        return expr
+
 
 class Backend(BasePandasBackend):
     name = 'pandas'
@@ -81,5 +109,35 @@ class Backend(BasePandasBackend):
     table_class = PandasTable
     client_class = PandasClient
 
-    def execute(self, *args, **kwargs):
-        return execute(*args, **kwargs)
+    def execute(self, query, params=None, limit='default', **kwargs):
+        if not hasattr(self, 'dictionary'):
+            return execute(query, params, limit, **kwargs)
+
+        if limit != 'default':
+            raise ValueError(
+                'limit parameter to execute is not yet implemented in the '
+                'pandas backend'
+            )
+
+        if not isinstance(query, ir.Expr):
+            raise TypeError(
+                "`query` has type {!r}, expected ibis.expr.types.Expr".format(
+                    type(query).__name__
+                )
+            )
+        return execute_and_reset(query, params=params, **kwargs)
+
+    def create_table(self, table_name, obj=None, schema=None):
+        """Create a table."""
+        if obj is None and schema is None:
+            raise com.IbisError('Must pass expr or schema')
+
+        if obj is not None:
+            df = pd.DataFrame(obj)
+        else:
+            dtypes = ibis_schema_to_pandas(schema)
+            df = schema.apply_to(
+                pd.DataFrame(columns=list(map(toolz.first, dtypes)))
+            )
+
+        self.dictionary[table_name] = df
